Dev Version Update(Almost ready for public version) -Added gases(Methane, Water Vapor, Oxygen and Hydrogen for now), simply liquids that are nearly invisible and fly instead of fall if possible -Fixed inconsistent fluids textures between item fluid containers and placed in world -Added Rubber Hanging Sign -Added LED Block, LED Panel, RGB Led Block and RGB Led Panel -Added Gas Infuser, nearly complete, but no restrictive fluid tanks for now -Updated to 1.20.1 -Added more recipes and changed some of them -Fixed Datagen in Forge -Added Basin Feature(basically a modified version of Lake Feature) -Added methane generation as a replacement of the air in a variation of rubber lakes(basin) -Added Armor Features, and then functions for Night Vision, Super Speed, Gravity 0 and Super Jump for electric armors items when charged and used the respective added key mapping

+++ common/src/main/java/wily/factocrafty/mixin/LivingEntityInjector.java
@@ -32,10 +32,10 @@ public abstract class LivingEntityInjector extends Entity {
         for (ItemStack i : getArmorSlots()) if (i.getItem() instanceof ArmorItem item && item.getMaterial().equals(FactocraftyArmorMaterials.RUBBER) && item.getEquipmentSlot() == EquipmentSlot.FEET) fallDistance /= 1.4;
     }
     @Inject(method = ("getJumpBoostPower"), at = @At(value = "RETURN"), cancellable = true)
-    private void injectJumpBoost(CallbackInfoReturnable<Double> info) {
+    private void injectJumpBoost(CallbackInfoReturnable<Float> info) {
         if ((Object)this instanceof Player p) {
             ItemStack s = p.getItemBySlot(EquipmentSlot.FEET);
-            if (s.getItem() instanceof ElectricArmorItem e && e.hasActiveFeature(ArmorFeatures.SUPER_JUMP,s,true)) info.setReturnValue(0.4+ info.getReturnValue());
+            if (s.getItem() instanceof ElectricArmorItem e && e.hasActiveFeature(ArmorFeatures.SUPER_JUMP,s,true)) info.setReturnValue(0.4F+ info.getReturnValue());
         }
     }
 }
